fixed another bad guy score bug

+++ pufferlib/ocean/blastar/blastar_env.h
@@ -31,7 +31,7 @@ typedef struct Log {
     float hit_enemy_with_bullet_rew;
     float hit_by_enemy_bullet_penalty_rew;
     int enemy_crossed_screen;
-    float bad_guy_score;
+    int bad_guy_score;
 } Log;
 
 // LogBuffer structure
@@ -101,7 +101,7 @@ typedef struct BlastarEnv {
     int bullet_travel_time;
     bool bullet_crossed_enemy_y; // Reset on bullet deactivation
     int kill_streak;
-    float bad_guy_score;
+    int bad_guy_score;
     Player player;
     Enemy enemy;               // Singular enemy
     Bullet bullet;
